pyproject.toml
[project]
name = "projecthub"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "black>=25.1.0",
    "botocore>=1.38.32",
    "django>=5.2.2",
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-storages>=1.14.6",
    "djangorestframework>=3.16.0",
    "factory-boy>=3.3.3",
    "psycopg2>=2.9.10",
    "pytest>=8.4.0",
    "pytest-django>=4.11.1",
    "pytest-mock>=3.14.1",
]
